pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["backend/agent.py", "backend/tools.py"]

[project]
name = "corporate-intelliops-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.10",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=1.1.11",
    "langsmith>=0.7.16",
    "markdown>=3.10.2",
    "httpx>=0.28.0",
    "markdownify>=1.2.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
    "langgraph-cli[inmem]>=0.1.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "streamlit>=1.55.0",
    "httpx-sse>=0.4.3",
    "xhtml2pdf>=0.2.16",
]
